// src/enums/category-code.enum.ts
export enum CategoryCodeEnum {
    FEATURE_MENU = 'FEATURE_MENU',
    FEATURE_OTHERS = 'FEATURE_OTHERS',
    EBOOK = 'EBOOK',
    EPUB_BOOK = 'EPUB_BOOK',
    PDF_BOOK = 'PDF_BOOK',
    // BOOK_CATEGORY
    BOOK_CATEGORY_FICTION = 'BOOK_CATEGORY_FICTION',
    BOOK_CATEGORY_NON_FICTION = 'BOOK_CATEGORY_NON_FICTION',
    BOOK_CATEGORY_CHILDREN = 'BOOK_CATEGORY_CHILDREN',
    BOOK_CATEGORY_TEEN = 'BOOK_CATEGORY_TEEN',
    BOOK_CATEGORY_ADULT = 'BOOK_CATEGORY_ADULT',
    BOOK_CATEGORY_OTHER = 'BOOK_CATEGORY_OTHER',
    // BOOK_STATUS
    BOOK_STATUS_PENDING = 'BOOK_STATUS_PENDING',
    BOOK_STATUS_APPROVED = 'BOOK_STATUS_APPROVED',
    BOOK_STATUS_REJECTED = 'BOOK_STATUS_REJECTED',
}